Refactor components.tokenizer
