+++ torchtitan/components/tokenizer.py
@@ -4,19 +4,13 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-import os
 
 from abc import ABC, abstractmethod
 
 
 class Tokenizer(ABC):
     # basic tokenizer interface, for typing purpose mainly
-    def __init__(self, tokenizer_path: str, from_pretrained: bool = False):
-        if not from_pretrained:
-            assert os.path.exists(
-                tokenizer_path
-            ), f"The tokenizer path does not exist: {tokenizer_path}"
-            assert os.path.isfile(tokenizer_path), tokenizer_path
+    def __init__(self):
         self._n_words = 8
 
     @abstractmethod
